Removed unused variable in misaligned.cpp

+++ misaligned.cpp
@@ -2,31 +2,30 @@
 #include <assert.h>
 #include <string.h>
 
-int getMaxLengthOfString(const char* colorList[]){
-    int maxLengthColor = 0;
+size_t getMaxLengthOfString(const char* colorList[]){
+    size_t maxLengthColor = 0;
     for(int i = 0; i < 5; i++)
         if(maxLengthColor < strlen(colorList[i]))
             maxLengthColor = strlen(colorList[i]);
     return maxLengthColor ;
 }
 
-std::string printWithSpacing(std::string colour, int length){
+std::string printWithSpacing(std::string colour, size_t length){
     std::string outputString = colour;
     size_t colourStringLength = colour.size();
-    int extraSpacesToBeAdded = length - colourStringLength;
+    size_t extraSpacesToBeAdded = length - colourStringLength;
     for(int i = 0; i <= extraSpacesToBeAdded; i++){
         outputString += " ";
     }
-
     return outputString;
 }
 
 int printColorMap() {
     const char* majorColor[] = {"White", "Red", "Black", "Yellow", "Violet"};
     const char* minorColor[] = {"Blue", "Orange", "Green", "Brown", "Slate"};
-    int maxLengthMajorColor = getMaxLengthOfString(majorColor);
-    int totalColourPair = sizeof(majorColor)/sizeof(majorColor[0]) * sizeof(minorColor)/sizeof(minorColor[0]);
-    int maxNumberOfDigits = std::to_string(totalColourPair).size();
+    size_t maxLengthMajorColor = getMaxLengthOfString(majorColor);
+    size_t totalColourPair = sizeof(majorColor)/sizeof(majorColor[0]) * sizeof(minorColor)/sizeof(minorColor[0]);
+    size_t maxNumberOfDigits = std::to_string(totalColourPair).size();
     int i = 0, j = 0;
     for(i = 0; i < 5; i++) {
         for(j = 0; j < 5; j++) {
